--- bakend/touchpoint/src/main/java/com/touchpoint/kh/config/WebConfig.java
@@ -1,17 +1,0 @@
-package com.touchpoint.kh.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**") // 모든 엔드포인트 허용
-                .allowedOrigins("http://localhost:4000") // React 앱의 주소
-                .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
-                .allowedHeaders("*") // 모든 헤더 허용
-                .allowCredentials(true); // 필요시 쿠키 허용
-    }
-}
