--- bakend/touchpoint/src/main/java/com/touchpoint/kh/user/config/ReactTest.java
@@ -1,20 +1,0 @@
-package com.touchpoint.kh.user.config;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api")
-public class ReactTest {
-
-
-	@CrossOrigin(origins = "http://localhost:3000")
-
-    @GetMapping("/test")
-    public ResponseEntity<String> testApi() {
-        return ResponseEntity.ok("Hello from Spring Boot!");
-    }
-}
