+++ frontend/src/styles/SignUp.css
@@ -12,7 +12,7 @@
     margin: 100px auto; /* 위아래 다시 설정해야함.*/
   }
 
-.form-container{
+.login-form-container{
     padding: auto;
     margin-bottom: 50px;
 }
